+++ changelog.md
@@ -1,12 +1,23 @@
 # MuditaOS changelog
 
+### Added
+* `[cellular]` SIM API for change pin, set/reset sim lock, switch sim card
+* `[notes]`    MMI/USSD communication window
+
 [Current release]
 
 ### Added
+* `[notes]` Notes application implemented.
+* `[notes]` Notes search engine implemented.
+
 
 ### Changed
 
+* `[desktop]` Windows refactor
+
 ### Fixed
+
+* `[meditation]` Fixed default counter settings 
 
 ### Other